requirements-test.txt
google-cloud-datastore==1.15.3
pytest==6.1.1
